refactor: rename authz package to auth
