+++ pkg/auth/handler.go
@@ -1,9 +1,9 @@
-package authz
+package auth
 
 import (
 	"net/http"
 
-	"github.com/eddycharly/generic-auth-server/pkg/authz/model"
+	"github.com/eddycharly/generic-auth-server/pkg/auth/model"
 )
 
 type Handler = func(*http.Request) *model.Response
