+++ pkg/auth/server.go
@@ -1,4 +1,4 @@
-package authz
+package auth
 
 import (
 	"context"
