+++ pkg/policy/compiler.go
@@ -4,10 +4,10 @@ import (
 	"errors"
 
 	"github.com/eddycharly/generic-auth-server/apis/v1alpha1"
-	engine "github.com/eddycharly/generic-auth-server/pkg/authz/cel"
-	"github.com/eddycharly/generic-auth-server/pkg/authz/cel/libs/auth"
-	"github.com/eddycharly/generic-auth-server/pkg/authz/cel/libs/http"
-	"github.com/eddycharly/generic-auth-server/pkg/authz/cel/utils"
+	engine "github.com/eddycharly/generic-auth-server/pkg/auth/cel"
+	"github.com/eddycharly/generic-auth-server/pkg/auth/cel/libs/auth"
+	"github.com/eddycharly/generic-auth-server/pkg/auth/cel/libs/http"
+	"github.com/eddycharly/generic-auth-server/pkg/auth/cel/utils"
 	"github.com/google/cel-go/cel"
 	"github.com/google/cel-go/common/types"
 	"github.com/google/cel-go/common/types/ref"
